mux: create separate path and wildcard types
Previously the path type acted as a wildcard and could only exist as a
leaf in the routing tree.
Now there are separate path and wild types, where wild behaves the same
way path did before and path is like string except that its value is the
remainder of the path (which still must be matched itself).